First Commit
This program generates a frequency response by outputting chosen sine waves and simultaneously recording the response. It is used to identify system response. In this form it was used to measure loudspeaker parameters. It can be modified to measure almost anything that can be sampled with a National Instruments DAQ.